First scene name change to Start, added BaseGame as game scene
We will use BaseGame to create our prefabs and then load them in as needed. Trying to implement state with JSON or maybe even text file loading of sorts might be introduced early to see if we can pull it off.

One note, please use the unity scripting API for these sorts of purposed, so we may be able to debug in the editor in case something goes horribly wrong.