this is how a property of an ingame object, like Attack could look like.
I have kept in mind that some characters, epic or not may not carry certain values, like attacking, etc. so i have decided to seperate them, so they can be mixed and matched as you like, keeping Ekaterinas comments in mind and to have sensible balancing decisions for all attributes in place.