the wolf sprite now carries a collider
I have added the wolf graphic into unity and it created a sprite render off it, i then resized the wolf sprite, to make it fit on a tile. I added a seperate empty game object and made it a child of the wolf sprite. This empty game object carries the box collider shape and has child parent relationship to wolf, awesome right?

Now its just a matter of implementing the old man and have a script be there that only makes the wolf move when the old man has already moved and attacked and we have turn based combat! Hope my descriptions help!

We just have to add the void OnTriggerEnter2D(Collider2D col) function as seen here: https://docs.unity3d.com/ScriptReference/Collider2D.OnTriggerEnter2D.html

and now we have the game object that collides with it and all the attached scripts and basically all the data we need to actually have combat going on!!!!!! WAAAAZAAAAAA! call me a magician, but you can apply this principle to all game elements really!