Added Wolf script and WolfAttackMove script
and offcourse also the ontriggerenter2d for all the necessary interaction, and triggering.

Remember, the enemy and characters are only allowed to attack once after moving, they have to move again, if they can in order to attack, if you cant attack a character than attack does not happen.

Just keep that in mind when thinking how are we going to figure out turn based combat.