WolfAttack should work when the collider touches the old man
I havent tested it.

we still need movement, but once that is in place, we still need to manage the amount of damage taken per round, the game needs to understand that

I assume we have to use some method inside of unity that lets the engine wait for player input, we may have disscussed this at one of the meetings, so the game kinda pauses and waits and the collider can not harm the old man infinitely.

// for now this will do, but there needs to be an empty gameobject that controlls the whole flow of everything
        // there will be a script attached to it, but this one will be accessed through a flag
        // so there will be only one instance of the round manager gameobject and by accessing the script you can get vital
        // information on the state of the round, like attack phase1 or 2 and move phase 1 or 2, or we will have it play out all at once and then repeat
        // still no clear approach, but yeah, there needs to be a central script where all the gameobjects can access game data about the phase 1 round is in.