added first steps in making limited movement possible for old man
first of all, sorry for not making lots of small commits.

i am just making it possible to set old mans position on the screen automatically through collision, setting that value to the old man script, accessing it in the mouse control script and i have added an empty, possibly private method that we will use to express the rules for movement.

Core question here is, how far is the old man able to move inside of the grid.

If we can determin that by comparing old mans current position on the grid, which is familiar and something we can messure, with the grid that has been clicked on, we can clearly set rules for all future characters of the squad that the player can control.

so a huge milestone has already been figured out, just by knowing, which grid the player has clicked on and on which grid a selected character has been standing on.

the rest is just comparing characters and integers, we just have not clearly stated yet how the old man is able to move.